add field for custom tokens
