+++ packages/augur-simplified/src/modules/modal/modal.styles.less
@@ -68,6 +68,10 @@
       margin-top: @size-24;
     }
 
+    > input {
+      margin-bottom: @size-16;
+    }
+
     > div[class*='ErrorBlock'] {
       margin-top: @size-8;
     }
